requirements.txt
attrs
twisted
emote-utils
autobahn
commandlet
